+++ src/libAtomVM/globalcontext.h
@@ -179,6 +179,10 @@ struct GlobalContext
     ErlNifResourceType *posix_dir_resource_type;
 #endif
 
+#ifdef HAVE_PSA_CRYPTO
+    ErlNifResourceType *psa_hash_op_resource_type;
+#endif
+
     void *platform_data;
 };